DiscordBot/requirements.txt
discord.py==2.3.1
yt-dlp==2025.3.23
python-dotenv==1.0.0
